Bug 2021180 - Make `HTMLEditor` not delete unnecessary preformatted linefeed when inserting content before it r=m_kato
Although for making the further processing simpler and faster, we always
delete the following invisible things when modifying the preceding
content. However, `www.icloud.com/notes` users the invisible
preformatted linefeed as anchor for the post-processing of inserting
text. Therefore, this patch makes the inserting content handlers stop
deleting preformatted linefeed. However, for minimizing the risk, this
does not make them stop deleting unnecessary `<br>` elements.

Note that different from padding `<br>` element, the invisible
preformatted linefeed appears as a line break in the flattened text
content for IME. Therefore, preserving the linefeed makes IME not
receive unexpected deletion of a line break. It might be better for IME.

Interestingly, the other browsers delete padding linefeed if it's a
padding for current empty block. It seems that they do it before
inserting content. However, this patch does not do that. Perhaps, we
should write a method to prepare to insert content and share it wider
in a follow up bug (bug 2021218).

Differential Revision: https://phabricator.services.mozilla.com/D286202

UltraBlame original commit: 7153b939ee3b22cede8a822221c5bab5e12ca284

+++ editor/libeditor/HTMLEditSubActionHandler.cpp
@@ -1137,7 +1137,12 @@ Result<EditActionResult, nsresult> HTMLEditor::HandleInsertText(
       
       insertEmptyTextResult.IgnoreCaretPointSuggestion();
       nsresult rv = EnsureNoFollowingUnnecessaryLineBreak(
-          insertEmptyTextResult.EndOfInsertedTextRef(), *editingHost);
+          insertEmptyTextResult.EndOfInsertedTextRef(),
+          
+          
+          
+          PreservePreformattedLineBreak::Yes, PaddingForEmptyBlock::Unnecessary,
+          *editingHost);
       if (NS_FAILED(rv)) {
         NS_WARNING(
             "HTMLEditor::EnsureNoFollowingUnnecessaryLineBreak() failed");
@@ -1328,25 +1333,34 @@ Result<EditActionResult, nsresult> HTMLEditor::HandleInsertText(
         if (nextThing.MaybeIgnoredLineBreak().isSome()) {
           const EditorLineBreak& lineBreak =
               nextThing.MaybeIgnoredLineBreak().ref();
-          const RefPtr<Element> ancestorLimiterToDeleteEmptyInlines =
-              lineBreak.ContentRef().IsInclusiveDescendantOf(
-                  currentPoint.GetContainer())
-                  ? currentPoint.GetContainerOrContainerParentElement()
-                  : editingHost.get();
-          {
-            AutoTrackDOMPoint trackCurrentPoint(RangeUpdaterRef(),
-                                                &currentPoint);
-            Result<EditorDOMPoint, nsresult> deleteLineBreakResultOrError =
-                DeleteLineBreakWithTransaction(
-                    lineBreak, nsIEditor::eStrip,
-                    *ancestorLimiterToDeleteEmptyInlines);
-            if (deleteLineBreakResultOrError.isErr()) [[unlikely]] {
-              NS_WARNING("HTMLEditor::DeleteLineBreakWithTransaction() failed");
-              return deleteLineBreakResultOrError.propagateErr();
+          
+          
+          
+          
+          
+          if (lineBreak.IsHTMLBRElement() ||
+              lineBreak.IsPaddingForEmptyBlock()) {
+            const RefPtr<Element> ancestorLimiterToDeleteEmptyInlines =
+                lineBreak.ContentRef().IsInclusiveDescendantOf(
+                    currentPoint.GetContainer())
+                    ? currentPoint.GetContainerOrContainerParentElement()
+                    : editingHost.get();
+            {
+              AutoTrackDOMPoint trackCurrentPoint(RangeUpdaterRef(),
+                                                  &currentPoint);
+              Result<EditorDOMPoint, nsresult> deleteLineBreakResultOrError =
+                  DeleteLineBreakWithTransaction(
+                      lineBreak, nsIEditor::eStrip,
+                      *ancestorLimiterToDeleteEmptyInlines);
+              if (deleteLineBreakResultOrError.isErr()) [[unlikely]] {
+                NS_WARNING(
+                    "HTMLEditor::DeleteLineBreakWithTransaction() failed");
+                return deleteLineBreakResultOrError.propagateErr();
+              }
             }
-          }
-          if (NS_WARN_IF(!currentPoint.IsSetAndValidInComposedDoc())) {
-            return Err(NS_ERROR_EDITOR_UNEXPECTED_DOM_TREE);
+            if (NS_WARN_IF(!currentPoint.IsSetAndValidInComposedDoc())) {
+              return Err(NS_ERROR_EDITOR_UNEXPECTED_DOM_TREE);
+            }
           }
         }
       }
@@ -1561,8 +1575,13 @@ Result<EditActionResult, nsresult> HTMLEditor::HandleInsertText(
         currentPoint = insertPaddingLineBreakResult.unwrap().UnwrapCaretPoint();
       }
     } else {
-      nsresult rv =
-          EnsureNoFollowingUnnecessaryLineBreak(currentPoint, *editingHost);
+      nsresult rv = EnsureNoFollowingUnnecessaryLineBreak(
+          currentPoint,
+          
+          
+          
+          PreservePreformattedLineBreak::Yes, PaddingForEmptyBlock::Unnecessary,
+          *editingHost);
       if (NS_FAILED(rv)) {
         NS_WARNING(
             "HTMLEditor::EnsureNoFollowingUnnecessaryLineBreak() failed");
@@ -2692,8 +2711,9 @@ HTMLEditor::DeleteTextAndNormalizeSurroundingWhiteSpaces(
   {
     AutoTrackDOMPoint trackPointToPutCaret(RangeUpdaterRef(),
                                            &newCaretPosition);
-    nsresult rv =
-        EnsureNoFollowingUnnecessaryLineBreak(newCaretPosition, aEditingHost);
+    nsresult rv = EnsureNoFollowingUnnecessaryLineBreak(
+        newCaretPosition, PreservePreformattedLineBreak::No,
+        PaddingForEmptyBlock::Significant, aEditingHost);
     if (NS_FAILED(rv)) {
       NS_WARNING("HTMLEditor::EnsureNoFollowingUnnecessaryLineBreak() failed");
       return Err(rv);
